requirements.txt
numpy==1.26.0
opencv-python==4.9.0.80
pyserial==3.5
qrcode==7.4.2
Pillow==10.2.0
ultralytics==8.0.0